Desktop/agentic AI project/DocuMind-RAG-Platform/DocuMind Enterprise Document RAG Backend/requirements.txt
fastapi
uvicorn
pydantic
requests
python-dotenv
pymupdf
langchain-text-splitters
chromadb
sentence-transformers
python-multipart
boto3
pinecone
